requirements.txt
# Core dependencies
aws-cdk-lib>=2.100.0
constructs>=10.0.0

# Agent framework
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# AgentCore SDK
bedrock-agentcore>=0.1.0

# AWS SDK
boto3>=1.34.0

# Chainlit UI
chainlit>=1.0.0

# Utilities
python-dotenv>=1.0.0
